Continuacion de vista
