+++ app/build.gradle
@@ -9,7 +9,7 @@ android {
 
     defaultConfig {
         applicationId "ayds.apolo"
-        minSdkVersion 16
+        minSdkVersion 24
         targetSdkVersion 33
         versionCode 1
         versionName "1.0"
