+++ app/src/main/java/ayds/newyork/songinfo/moredetails/injector/MoreDetailsDependenciesInjector.kt
@@ -7,18 +7,18 @@ import ayds.newyork.songinfo.moredetails.data.local.sqldb.CursorToArtistInfoMapp
 import ayds.newyork.songinfo.moredetails.domain.DataRepository
 import ayds.newyork.songinfo.moredetails.presentation.presenter.MoreDetailsPresenter
 import ayds.newyork.songinfo.moredetails.presentation.presenter.MoreDetailsPresenterImpl
-import ayds.newyork.songinfo.moredetails.presentation.view.FormatterInfo
 import ayds.newyork.songinfo.moredetails.presentation.view.OtherInfoViewActivity
 import ayds.newyork.songinfo.moredetails.data.BrokerServiceImpl
 import ayds.newyork.songinfo.moredetails.data.Proxy
 import ayds.newyork.songinfo.moredetails.data.proxy.NYTProxyImpl
 import ayds.newyork.songinfo.moredetails.data.proxy.WikipediaProxyImpl
 import ayds.newyork.songinfo.moredetails.data.proxy.LastFMProxyImpl
+import ayds.newyork.songinfo.moredetails.presentation.view.FormatterInfo
 
 object MoreDetailsDependenciesInjector {
+    /*private const val NYTIMES_URL = "https://api.nytimes.com/svc/search/v2/"
+    val otherInfo: OtherInfoViewActivity = OtherInfoViewActivity(formatterInfo = FormatterInfo(""))*/
     private lateinit var moreDetailsPresenter: MoreDetailsPresenter
-    private const val NYTIMES_URL = "https://api.nytimes.com/svc/search/v2/"
-    val otherInfo: OtherInfoViewActivity = OtherInfoViewActivity(formatterInfo = FormatterInfo(""))
     private lateinit var broker: BrokerServiceImpl
     private var proxys: MutableList<Proxy> = mutableListOf()
     fun init(otherInfoView: OtherInfoViewActivity) {
