+++ app/src/main/java/ayds/newyork/songinfo/moredetails/presentation/view/MoreDetailsViewConstants.kt
@@ -4,8 +4,6 @@ const val NO_RESULTS = "No Results"
 const val HTML_DIV_WIDTH = "<html><div width=400>"
 const val HTML_FONT_FACE = "<font face=\"arial\">"
 const val HTML_CLOSE_TAGS = "</font></div></html>"
-const val APOSTROPHE = "'"
-const val SPACE = " "
 const val ENTER_LINE = "\n"
 const val ENTER_LINE_ESCAPE_SEQ = "\\n"
 const val HTML_BREAK = "<br>"
