+++ app/src/main/res/layout/activity_main.xml
@@ -13,7 +13,7 @@
         android:layout_margin="16dp"
         android:hint="@string/song_name"
         android:importantForAutofill="no"
-        app:layout_constraintBottom_toTopOf="@+id/descriptionTextView"
+        app:layout_constraintBottom_toTopOf="@+id/description"
         app:layout_constraintEnd_toStartOf="@+id/searchButton"
         app:layout_constraintHorizontal_chainStyle="packed"
         app:layout_constraintStart_toStartOf="parent"
@@ -30,7 +30,7 @@
         app:layout_constraintTop_toTopOf="parent" />
 
     <TextView
-        android:id="@id/descriptionTextView"
+        android:id="@id/description"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:padding="16dp"
@@ -49,7 +49,7 @@
         android:text="@string/more_details"
         app:layout_constraintEnd_toStartOf="@+id/openSongButton"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@id/descriptionTextView" />
+        app:layout_constraintTop_toBottomOf="@id/description" />
 
     <Button
         android:id="@id/openSongButton"
@@ -60,7 +60,7 @@
         android:text="@string/open_song"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toEndOf="@id/modeDetailsButton"
-        app:layout_constraintTop_toBottomOf="@id/descriptionTextView" />
+        app:layout_constraintTop_toBottomOf="@id/description" />
 
     <ImageView
         android:id="@+id/posterImageView"
